List of Functions

Single Reactor Solvers

ADM1jl.ADM1sol — Function

function ADM1sol(tspan::Tuple,u0::Vector,IV::Vector; <keyword arguments>)

Compute the solution for the given timespan, tspan; initial condition, u0; and inflow vector IV.

Also return the time (in seconds) the solution took to compute. The difference between this function and ExampleSol is that this function reads in the parameter values from a .csv file.

Optional Arguments

alg = Rodas4P(): the ODE solver algorithm.

tols = 1e-4: the absolute and relative tolerance of the solver method.

tMax = 300.0: the maximum time (in seconds), that the function will run before timing out.

saveAt = []: specific times to save the solution. If given a number n, the solver will save the solution every n timesteps

Examples

julia> u0 = initialConditions();

julia> IV = inflowvector_definition();

julia> sol, tSol = ADM1sol((0.0,200.0),u0,IV);

julia> sol
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 146-element Vector{Float64}:
[...]

u: 146-element Vector{Vector{Float64}}:
[...]

julia> tSol
0.3854937

function ADM1sol(tspan::Tuple,u0::Vector,IV::Vector{Vector{Float64}},IVtimes::Vector{Float64}; <keyword arguments>)

Compute the solution with variable inflow for the given timespan, tspan; initial condition, u0; and variable inflow vector IV where each entry in IV corresponds to the inflow vector at the corresponding entry in IVtimes

Also return the time (in seconds) the solution took to compute.

Optional Arguments

alg = Rodas4P(): the ODE solver algorithm.

tols = 1e-4: the absolute and relative tolerance of the solver method.

tMax = 300.0: the maximum time (in seconds), that the function will run before timing out.

saveAt = []: specific times to save the solution. If given a number n, the solver will save the solution every n timesteps

Examples

julia> u0 = initialConditions();

julia> t = [i for i in 0.0:0.1:50.0];

julia> IV_temp = inflowvector_definition();

julia> IV = [IV_temp*(0.5*rand()+1.0) for i in 1:length(t)]

julia> sol,tSol = ADM1sol((0.0,50.0),u0,IV,t);

julia> sol
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 2011-element Vector{Float64}:
[...]

u: 2011-element Vector{Vector{Float64}}:
[...]

julia> tSol
14.4643811

Multiple Reactor Solvers

ADM1jl.MultiChamberSolution — Function

function MultiChamberSolution(tspan::Tuple,u0::Tuple,IV::Vector,nChambers::Int64; <keyword arguments>)

Compute the solution for a system of nChambers connected CSTRs with u0 initial conditions. The inflow of the first CSTR is given by IV, the outflow of the first CSTR becomes the inflow of second CSTR and so on.

This function reads in the parameter values from a .csv file. The names of the .csv files that contain the parameter values should be "modelparameters.csv", "modelparameters2.csv", "model_parameters3.csv", ... and so on.

Arguments

alg = Rodas4P(): the ODE solver algorithm.

tols = 1e-4: the absolute and relative tolerance of the solver method.

tMax = 300.0: the maximum time (in seconds), that the function will run before timing out.

saveAt = []: specific times to save the solution. If given a number n, the solver will save the solution every n timesteps

Examples

julia> u0 = initialConditions();

julia> IV = inflowvector_definition();

julia> sols = MultiChamberSolution((0.0,200.0),(u0,u0,u0),IV,3);
Finished Chamber 1
Finished Chamber 2
Finished Chamber 3

julia> sols[1]
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 115-element Vector{Float64}:
[...]

u: 115-element Vector{Vector{Float64}}:
[...]

julia> sols[2]
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 115-element Vector{Float64}:
[...]

u: 115-element Vector{Vector{Float64}}:
[...]

julia> sols[3]
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 115-element Vector{Float64}:
[...]

u: 115-element Vector{Vector{Float64}}:
[...]

function MultiChamberSolution(tspan::Tuple,u0::Tuple,IV::Vector{Vector{Float64}},IVtimes::Vector{Float64},nChambers::Int64;<keyword arguments>)

Compute the solution for a system of nChambers connected CSTRs with u0 initial conditions. The inflow of the first CSTR is given by variable inflow vector IV where each entry in IV corresponds to the inflow vector at the corresponding entry in IVtimes, the outflow of the first CSTR becomes the inflow of second CSTR and so on.

This function reads in the parameter values from a .csv file. The names of the .csv files that contain the parameter values should be "modelparameters.csv", "modelparameters2.csv", "model_parameters3.csv", ... and so on.

Arguments

alg = Rodas4P(): the ODE solver algorithm.

tols = 1e-4: the absolute and relative tolerance of the solver method.

tMax = 300.0: the maximum time (in seconds), that the function will run before timing out.

saveAt = []: specific times to save the solution. If given a number n, the solver will save the solution every n timesteps

Examples

julia> u0 = initialConditions();

julia> t = [i for i in 0.0:0.1:50.0];

julia> IV_temp = inflowvector_definition();

julia> IV = [IV_temp*(0.5*rand()+1.0) for i in 1:length(t)];

julia> sols = MultiChamberSolution((0.0,50.0),(u0,u0,u0),IV,t,3);
Finished Chamber 1
Finished Chamber 2
Finished Chamber 3

julia> sols[1]
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 115-element Vector{Float64}:
[...]

u: 115-element Vector{Vector{Float64}}:
[...]

julia> sols[2]
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 115-element Vector{Float64}:
[...]

u: 115-element Vector{Vector{Float64}}:
[...]

julia> sols[3]
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 115-element Vector{Float64}:
[...]

u: 115-element Vector{Vector{Float64}}:
[...]

Other Functions

ADM1jl.RHSfun — Method

RHSfun(du,u,p,t)

Return the right-hand side of the system of ODEs, this is an in-place function.

Arguments

du::Vector: the rate change of the state vector (required since the function is defined in-place).

u::Vector: the state vector.

p::Vector: all of the model parameters.

t: the timestep, usually a Float64.

ADM1jl.reactionrates — Method

reactionrates(bp,rp,php,pressures,sx,NREAC::Int)

Compute and return the vector of reaction rates.

Arguments

bp::Vector: the biochemical parameters.

rp::Vector: the reactor parameters.

php::Vector: the physiochemical parameters.

pressures::Vector: the gas pressures.

sx::Vector: the state vector.

NREAC::Integer: the number of reaction rates.

Examples

julia> u0 = initialConditions();

julia> bp = biochemicalparameter_definition();

julia> rp = reactorParameterDefinition();

julia> php = physiochemicalParameterDefinition(rp);

julia> pressures = pressureOfGasses(u0,php,rp);

julia> NREAC = 29;

julia> reactionrates(bp,rp,php,pressures,u0,NREAC)
29-element Vector{Real}:
 0.155
 0.28
 1.0
 0.29000000000000004
 0.2950855111452082
 ⋮
 0.0
 7.402547081085608e-6
 1.295220255437568
 0.052518812965057435

ADM1jl.pressureOfGasses — Method

pressureOfGasses(sx,php,rp)

Compute the pressures of the gasses.

Arguments

sx::Vector: the state vector.

php::Vector: the physicochemical parameters.

rp::Vector: the reactor parameters.

Examples

julia> u0 = initialConditions();

julia> rp = reactorParameterDefinition();

julia> php = physiochemicalParameterDefinition(rp);

julia> pressureOfGasses(u0,php,rp)
5-element Vector{Float64}:
    1.6333471490625e-5
    0.6525381992578124
    0.35869584449999997
    1.0669181223042932
    2695.9061152146637

ADM1jl.monod — Method

monod(u, k)

Compute the monod function u/(u+k) where u is the state varible and k is the half-saturation concentration.

If u <= 0 return 0.

Examples

julia> monod(3.0,2.0) # when u is non-zero positive
0.6

julia> monod(-3.0,2.0) # when u is negative
0