List of Functions

Functions

ADM1jl.ADM1sol — Method

function ADM1sol(tspan::Tuple,u0::Vector,IV::Vector; <keyword arguments>)

Compute the solution for the given timespan, tspan; initial condition, u0; and inflow vector IV.

Also return the time (in seconds) the solution took to compute. The difference between this function and ExampleSol is that this function reads in the parameter values from a .csv file.

Optional Arguments

alg = Rodas4P(): the ODE solver algorithm.

tols = 1e-4: the absolute and relative tolerance of the solver method.

tMax = 300.0: the maximum time (in seconds), that the function will run before timing out.

Examples

julia> u0 = ADM1jl.InitialConditions();

julia> IV = ADM1jl.inflowvector_definition();

julia> sol, tSol = ADM1jl.ADM1sol((0.0,200.0),u0,IV);

julia> sol
retcode: Success
Interpolation: specialized 3rd order "free" stiffness-aware interpolation
t: 146-element Vector{Float64}:
[...]

u: 146-element Vector{Vector{Float64}}:
[...]

julia> tSol
0.3854937

ADM1jl.RHSfun — Method

RHSfun(du,u,p,t)

Return the right-hand side of the system of ODEs, this is an in-place function.

Arguments

du::Vector: the rate change of the state vector (required since the function is defined in-place).

u::Vector: the state vector.

p::Vector: all of the model parameters.

t: the timestep, usually a Float64.

ADM1jl.reactionrates — Method

reactionrates(bp,rp,php,pressures,sx,NREAC::Int)

Compute and return the vector of reaction rates.

Arguments

bp::Vector: the biochemical parameters.

rp::Vector: the reactor parameters.

php::Vector: the physiochemical parameters.

pressures::Vector: the gas pressures.

sx::Vector: the state vector.

NREAC::Integer: the number of reaction rates.

Examples

julia> u0 = ADM1jl.InitialConditions();

julia> bp = ADM1jl.biochemicalparameter_definition();

julia> rp = ADM1jl.reactorParameterDefinition();

julia> php = ADM1jl.physiochemicalParameterDefinition(rp);

julia> pressures = ADM1jl.pressureOfGasses(u0,php,rp);

julia> NREAC = 29;

julia> ADM1jl.reactionrates(bp,rp,php,pressures,u0,NREAC)
29-element Vector{Real}:
 0.155
 0.28
 1.0
 0.29000000000000004
 0.2950855111452082
 ⋮
 0.0
 7.402547081085608e-6
 1.295220255437568
 0.052518812965057435

ADM1jl.pressureOfGasses — Method

pressureOfGasses(sx,php,rp)

Compute the pressures of the gasses.

Arguments

sx::Vector: the state vector.

php::Vector: the physicochemical parameters.

rp::Vector: the reactor parameters.

Examples

julia> u0 = ADM1jl.InitialConditions();

julia> rp = ADM1jl.reactorParameterDefinition();

julia> php = ADM1jl.physiochemicalParameterDefinition(rp);

julia> ADM1jl.pressureOfGasses(u0,php,rp)
5-element Vector{Float64}:
    1.6333471490625e-5
    0.6525381992578124
    0.35869584449999997
    1.0669181223042932
 2695.9061152146637

ADM1jl.monod — Method

monod(u, k)

Compute the monod function u/(u+k) where u is the state varible and k is the half-saturation concentration.

If u <= 0 return 0.

Examples

julia> ADM1jl.monod(3.0,2.0) # when u is non-zero positive
0.6

julia> ADM1jl.monod(-3.0,2.0) # when u is negative
0